+++ xbmc/cores/VideoPlayer/DVDStreamInfo.h
@@ -62,7 +62,7 @@ public:
   int fpsrate;
   int height; // height of the stream reported by the demuxer
   int width; // width of the stream reported by the demuxer
-  float aspect; // display aspect as reported by demuxer
+  double aspect; // display aspect as reported by demuxer
   bool vfr; // variable framerate
   bool stills; // there may be odd still frames in video
   int level; // encoder level of the stream reported by the decoder. used to qualify hw decoders.
